Original change: https://android-review.googlesource.com/c/platform/external/perfetto/+/1315194

+++ src/trace_processor/tables/slice_tables.h
@@ -83,7 +83,7 @@ PERFETTO_TP_TABLE(PERFETTO_TP_SCHED_SLICE_TABLE_DEF);
   C(StringPool::Id, command_buffer_name)            \
   C(base::Optional<uint32_t>, frame_id)             \
   C(base::Optional<uint32_t>, submission_id)        \
-  C(base::Optional<uint32_t>, hw_queue_id)
+  C(base::Optional<int64_t>, hw_queue_id)
 
 PERFETTO_TP_TABLE(PERFETTO_TP_GPU_SLICES_DEF);
